+++ g_items.c
@@ -1298,10 +1298,10 @@ and for each item in each client's inventory.
 */
 void PrecacheItem (gitem_t *it)
 {
-	char	*s, *start;
-	char	data[MAX_QPATH];
-	int		len;
-	gitem_t	*ammo;
+	char		*s, *start;
+	char		data[MAX_QPATH];
+	intptr_t	len;
+	gitem_t		*ammo;
 
 	if (!it)
 		return;
