+++ g_monster.c
@@ -14,7 +14,7 @@ void monster_fire_bullet (edict_t *self, vec3_t start, vec3_t dir, int damage, i
 	fire_bullet (self, start, dir, damage, kick, hspread, vspread, MOD_UNKNOWN);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }
@@ -24,7 +24,7 @@ void monster_fire_shotgun (edict_t *self, vec3_t start, vec3_t aimdir, int damag
 	fire_shotgun (self, start, aimdir, damage, kick, hspread, vspread, count, MOD_UNKNOWN);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }
@@ -34,7 +34,7 @@ void monster_fire_blaster (edict_t *self, vec3_t start, vec3_t dir, int damage, 
 	fire_blaster (self, start, dir, damage, speed, effect, false);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }	
@@ -45,7 +45,7 @@ void monster_fire_blaster2 (edict_t *self, vec3_t start, vec3_t dir, int damage,
 	fire_blaster2 (self, start, dir, damage, speed, effect, false);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }
@@ -56,7 +56,7 @@ void monster_fire_tracker (edict_t *self, vec3_t start, vec3_t dir, int damage, 
 	fire_tracker (self, start, dir, damage, speed, enemy);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }
@@ -66,7 +66,7 @@ void monster_fire_heat (edict_t *self, vec3_t start, vec3_t dir, vec3_t offset, 
 	fire_heat (self, start, dir, offset, damage, kick, true);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }
@@ -77,7 +77,7 @@ void monster_fire_grenade (edict_t *self, vec3_t start, vec3_t aimdir, int damag
 	fire_grenade (self, start, aimdir, damage, speed, 2.5, damage+40);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }
@@ -87,7 +87,7 @@ void monster_fire_rocket (edict_t *self, vec3_t start, vec3_t dir, int damage, i
 	fire_rocket (self, start, dir, damage, speed, damage+20, damage);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }	
@@ -99,7 +99,7 @@ void monster_fire_railgun (edict_t *self, vec3_t start, vec3_t aimdir, int damag
 		fire_rail (self, start, aimdir, damage, kick);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }
@@ -109,7 +109,7 @@ void monster_fire_bfg (edict_t *self, vec3_t start, vec3_t aimdir, int damage, i
 	fire_bfg (self, start, aimdir, damage, speed, damage_radius);
 
 	gi.WriteByte (svc_muzzleflash2);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WriteByte (flashtype);
 	gi.multicast (start, MULTICAST_PVS);
 }
